+++ jobs/__init__.py
@@ -1,0 +1,19 @@
+import os
+import logging
+from pathlib import Path
+
+
+ENV = os.getenv("ENV", "dev")
+BASE = Path(__file__).parent
+
+
+
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.INFO)
+
+if not logger.hasHandlers():
+    console = logging.StreamHandler()
+    console.setLevel(logging.INFO)
+    fmt = logging.Formatter("%(asctime)s [%(levelname)s] %(name)s: %(message)s")
+    console.setFormatter(fmt)
+    logger.addHandler(console)
