+++ jobs/data/2023_box_office_number_ones.txt
@@ -1,0 +1,35 @@
+# https://en.wikipedia.org/wiki/List_of_2023_box_office_number-one_films_in_the_United_States
+
+Avatar:_The_Way_of_Water
+Knock_at_the_Cabin
+Magic_Mike's_Last_Dance
+Ant-Man_and_the_Wasp:_Quantumania
+Creed_III
+Scream_VI
+Shazam!_Fury_of_the_Gods
+John_Wick:_Chapter_4
+Dungeons_&_Dragons:_Honor_Among_Thieves
+The_Super_Mario_Bros._Movie
+Guardians_of_the_Galaxy_Vol._3
+Fast_X
+The_Little_Mermaid_(2023_film)
+Spider-Man:_Across_the_Spider-Verse
+Transformers:_Rise_of_the_Beasts
+The_Flash_(film)
+Indiana_Jones_and_the_Dial_of_Destiny
+Insidious:_The_Red_Door
+Mission:_Impossible_–_Dead_Reckoning_Part_One
+Barbie_(film)
+Blue_Beetle_(film)
+Gran_Turismo_(film)
+The_Equalizer_3
+The_Nun_II
+PAW_Patrol:_The_Mighty_Movie
+The_Exorcist:_Believer
+Five_Nights_at_Freddy's_(film)
+The_Marvels
+The_Hunger_Games:_The_Ballad_of_Songbirds_&_Snakes
+Renaissance:_A_Film_by_Beyoncé
+The_Boy_and_the_Heron
+Wonka_(film)
+Aquaman_and_the_Lost_Kingdom
