+++ jobs/data/2024_box_office_number_ones.txt
@@ -1,0 +1,20 @@
+# https://en.wikipedia.org/wiki/List_of_2024_box_office_number-one_films_in_the_United_States
+
+Wonka_(film)
+Mean_Girls_(2024_film)
+Argylle
+Bob_Marley:_One_Love
+Dune:_Part_Two
+Kung_Fu_Panda_4
+Ghostbusters:_Frozen_Empire
+Godzilla_x_Kong:_The_New_Empire
+Civil_War_(film)
+Challengers_(film)
+The_Fall_Guy_(2024_film)
+Kingdom_of_the_Planet_of_the_Apes
+IF_(film)
+Furiosa:_A_Mad_Max_Saga
+The_Garfield_Movie
+Bad_Boys:_Ride_or_Die
+Inside_Out_2
+Despicable_Me_4
