+++ jobs/data/best_selling.txt
@@ -1,0 +1,22 @@
+# https://en.wikipedia.org/wiki/List_of_best-selling_films_in_the_United_States
+
+Shrek_2
+How_the_Grinch_Stole_Christmas_(2000_film)
+One_Hundred_and_One_Dalmatians
+The_Dark_Knight
+Frozen_(2013_film)
+Cars_(film)
+Independence_Day_(1996_film)
+Shrek
+Cinderella_(1950_film)
+Jurassic_Park_(film)
+Spider-Man_(2002_film)
+Avatar_(2009_film)
+Toy_Story
+Titanic_(1997_film)
+Beauty_and_the_Beast_(1991_film)
+Aladdin_(1992_Disney_film)
+Snow_White_and_the_Seven_Dwarfs_(1937_film)
+The_Little_Mermaid_(1989_film)
+Finding_Nemo
+The_Lion_King
