+++ jobs/data/imdb_250.txt
@@ -1,0 +1,252 @@
+# https://www.imdb.com/list/ls505855487/
+
+12_Angry_Men_(1957_film)
+12_Years_a_Slave_(film)
+1917_(2019_film)
+2001:_A_Space_Odyssey_(film)
+3_Idiots
+A_Separation
+A_Beautiful_Mind_(film)
+A_Clockwork_Orange_(film)
+Aladdin_(1992_Disney_film)
+Alien_(film)
+Aliens_(film)
+All_About_Eve
+Amadeus_(film)
+Amélie
+American_History_X
+American_Beauty_(1999_film)
+Amores_Perros
+Apocalypse_Now
+Avengers:_Endgame
+Avengers:_Infinity_War
+Back_to_the_Future
+Barry_Lyndon
+Batman_Begins
+Before_Sunrise
+Before_Sunset
+Ben-Hur_(1959_film)
+Bicycle_Thieves
+Blade_Runner
+Braveheart
+Capernaum_(film)
+Casablanca_(film)
+Casino_(1995_film)
+Catch_Me_If_You_Can
+Children_of_Heaven
+Chinatown_(1974_film)
+Cinema_Paradiso
+Citizen_Kane
+City_Lights
+City_of_God_(2002_film)
+Coco_(2017_film)
+Come_and_See
+Cool_Hand_Luke
+Dances_with_Wolves
+Dangal_(film)
+Das_Boot_(film)
+Dead_Poets_Society
+Detective_Pikachu_(film)
+Dial_M_for_Murder
+Die_Hard
+Django_Unchained
+Double_Indemnity
+Downfall_(2004_film)
+Dr._Strangelove_or:_How_I_Learned_to_Stop_Worrying_and_Love_the_Bomb
+Eternal_Sunshine_of_the_Spotless_Mind
+Fargo_(1996_film)
+Father_(2020_film)
+Fight_Club
+Finding_Nemo
+For_a_Few_Dollars_More
+Ford_v_Ferrari
+Forrest_Gump
+Full_Metal_Jacket
+Gladiator_(2000_film)
+Gone_Girl_(film)
+Gone_with_the_Wind_(film)
+Good_Will_Hunting
+Goodfellas
+Gran_Torino
+Grave_of_the_Fireflies
+Green_Book_(film)
+Groundhog_Day_(film)
+Guardians_of_the_Galaxy_Vol._3
+Hachi:_A_Dog%27s_Tale
+Hacksaw_Ridge
+Hamilton_(2020_film)
+Harakiri_(1962_film)
+Harry_Potter_and_the_Deathly_Hallows:_Part_2
+Heat_(1995_film)
+High_and_Low_(1963_film)
+Hotel_Rwanda
+How_to_Train_Your_Dragon_(film)
+Howl%27s_Moving_Castle_(film)
+Ikiru
+In_the_Name_of_the_Father_(film)
+Incendies
+Inception
+Indiana_Jones_and_the_Last_Crusade
+Indiana_Jones_and_the_Raiders_of_the_Lost_Ark
+Inglourious_Basterds
+Inside_Out_(2015_film)
+Interstellar_(film)
+Into_the_Wild_(film)
+It_Happened_One_Night
+It%27s_a_Wonderful_Life
+Jai_Bhim_(film)
+Jaws_(film)
+Joker_(2019_film)
+Judgment_at_Nuremberg
+Jurassic_Park_(film)
+Kill_Bill:_Vol._1
+Klaus_(film)
+L.A._Confidential_(film)
+La_haine
+Lawrence_of_Arabia_(film)
+Léon:_The_Professional
+Life_Is_Beautiful
+Life_of_Brian
+Like_Stars_on_Earth
+Lock,_Stock_and_Two_Smoking_Barrels
+Logan_(film)
+M_(1931_film)
+Mad_Max:_Fury_Road
+Mary_and_Max
+Memento_(film)
+Memories_of_Murder
+Metropolis_(1927_film)
+Million_Dollar_Baby
+Modern_Times_(film)
+Monsters,_Inc.
+Monty_Python_and_the_Holy_Grail
+Mr._Smith_Goes_to_Washington
+My_Father_and_My_Son
+My_Neighbor_Totoro
+Network_(1976_film)
+No_Country_for_Old_Men
+North_by_Northwest
+Oldboy_(2003_film)
+On_the_Waterfront
+Once_Upon_a_Time_in_America
+Once_Upon_a_Time_in_the_West
+One_Flew_Over_the_Cuckoo%27s_Nest_(film)
+Pan%27s_Labyrinth
+Parasite_(2019_film)
+Pather_Panchali
+Paths_of_Glory
+Persona_(1966_film)
+Pirates_of_the_Caribbean:_The_Curse_of_the_Black_Pearl
+Platoon_(film)
+Princess_Mononoke
+Prisoners_(2013_film)
+Psycho_(1960_film)
+Pulp_Fiction
+Raging_Bull
+Ran_(film)
+Rashomon
+Ratatouille_(film)
+Rear_Window
+Rebecca_(1940_film)
+Requiem_for_a_Dream
+Reservoir_Dogs
+Rocky
+Room_(2015_film)
+Rush_(2013_film)
+Saving_Private_Ryan
+Scarface_(1983_film)
+Schindler%27s_List
+Seven_Samurai
+Seven_(1995_film)
+Sherlock_Jr.
+Shutter_Island
+Singin%27_in_the_Rain
+Snatch_(film)
+Some_Like_It_Hot
+Spider-Man:_Across_the_Spider-Verse
+Spider-Man:_Into_the_Spider-Verse
+Spider-Man:_No_Way_Home
+Spirited_Away
+Spotlight_(film)
+Stand_by_Me_(film)
+Star_Wars:_Episode_IV_-_A_New_Hope
+Star_Wars:_Episode_VI_-_Return_of_the_Jedi
+Sunset_Boulevard_(film)
+Taxi_Driver
+Terminator_2:_Judgment_Day
+The_400_Blows
+The_Apartment
+The_Best_Years_of_Our_Lives
+The_Big_Lebowski
+The_Bridge_on_the_River_Kwai
+The_Dark_Knight
+The_Dark_Knight_Rises
+The_Deer_Hunter
+The_Departed
+The_Exorcist
+The_Godfather
+The_Godfather_Part_II
+The_Gold_Rush
+The_Good,_the_Bad_and_the_Ugly
+The_Grand_Budapest_Hotel
+The_Grapes_of_Wrath_(film)
+The_Great_Dictator
+The_Handmaiden
+The_Incredibles
+The_Intouchables
+The_Iron_Giant
+The_Lion_King
+The_Lives_of_Others
+The_Lord_of_the_Rings:_The_Fellowship_of_the_Ring
+The_Lord_of_the_Rings:_The_Return_of_the_King
+The_Lord_of_the_Rings:_The_Two_Towers
+The_Matrix
+The_Passion_of_Joan_of_Arc
+The_Prestige_(film)
+The_Secret_in_Their_Eyes
+The_Seventh_Seal
+The_Shawshank_Redemption
+The_Sixth_Sense
+The_Sting
+The_Terminator
+The_Third_Man
+The_Truman_Show
+The_Usual_Suspects
+The_Wages_of_Fear
+The_Battle_of_Algiers
+The_Elephant_Man_(film)
+The_Empire_Strikes_Back
+The_General_(1926_film)
+The_Great_Escape_(film)
+The_Green_Mile_(film)
+The_Help_(film)
+The_Hunt_(2012_film)
+The_Kid_(1921_film)
+The_Last_Samurai
+The_Pianist_(2002_film)
+The_Shining_(film)
+The_Silence_of_the_Lambs_(film)
+The_Sound_of_Music_(film)
+The_Thing_(1982_film)
+The_Treasure_of_the_Sierra_Madre_(film)
+The_Wizard_of_Oz_(1939_film)
+The_Wolf_of_Wall_Street_(2013_film)
+There_Will_Be_Blood
+Three_Billboards_Outside_Ebbing,_Missouri
+To_Be_or_Not_to_Be_(1942_film)
+To_Kill_a_Mockingbird_(film)
+Tokyo_Story
+Top_Gun:_Maverick
+Toy_Story
+Toy_Story_3
+Trainspotting_(film)
+Unforgiven
+Up_(2009_film)
+V_for_Vendetta
+Vertigo_(film)
+WALL-E
+Warrior_(2011_film)
+Whiplash_(2014_film)
+Wild_Strawberries_(film)
+Witness_for_the_Prosecution_(1957_film)
+Yojimbo
