+++ jobs/data/portrait_prompts.txt
@@ -1,0 +1,81 @@
+actor;A professional headshot of a young, handsome actor
+actor;A professional headshot of an old, well respected, actor
+actor;An actor on the red carpet of an award gala
+actor;An old comedian performing standup
+actor;A classical English theater actor
+actor;An actor on the set of a comedy film
+actor;Actor from an 1800s costume drama
+actor;A Victorian actor performing a Shakespeare play
+actor;An actor passionately delivering a monologue
+actor;Comedic Actor in a Slapstick Scene
+actor;Action Movie Star in a Fight Sequence
+actor;An actor in an intense fight scene
+actor;Romantic Lead in a Rainy Scene
+actor;Sci-Fi Actor in an Alien Encounter
+actor;An actor dressed as a historical epic warrior
+actor;An actor spills coffee on another actor in a bustling café. Their eyes lock, and a smile forms
+actor;An actor with half-human, half-metallic features gazes into the distance, contemplating their existence
+actor;An actor twirls in a vibrant dance sequence
+actor;An actor, wide-eyed and terrified, runs through a dark forest, pursued by an unseen menace
+actor;Voice actor for an animated TV show
+actor;An actor records lines in a soundproof booth, bringing a beloved animated character to life
+actor;An actor, impeccably dressed, exchanges coded messages with another actor
+actor;An actor answering questions in a panel
+actor;A lead actor in the premiere of a movie
+actor;A professional headshot of an actor, capturing their essence and charisma
+actor;An actor looks directly into the camera, exuding confidence and charm
+actor;An image of the actor at a glamorous film premiere, dressed in designer attire
+actor;A still from a televised awards ceremony, showing an actor holding a trophy and delivering a heartfelt speech
+actor;A behind-the-scenes photo of an actor on a film set, immersed in their role
+actor;A side-by-side comparison of an actor in and out of character
+actor;An actor recreating a famous scene from one of their acclaimed movies
+actor;An actor captured mid-performance during a theater play
+actor;An actor performing a Shakespeare play on stage
+actor;An image of an actor participating in a humanitarian cause
+actor;A candid shot of an actor sharing a laugh with the film crew
+actor;An actor striking a captivating pose on a movie poster
+actor;An actor wearing a tuxedo in being interviewed
+actor;A behind the scenes shot of an actor within the a set on a sound stage
+actor;An actor next to a microphone answering questions
+actor;A full body portrait of an actor
+actor;An actor with their family in a photo op
+actor;An actor performing a musical number on a Broadway show
+actor;An actor participating in a panel discussion
+actor;An actor hosting a fancy gala event
+actor;An actor looking over their shoulder while being photographed
+actor;A well respected comedian walking down the street
+actor;A classic black-and-white portrait of an elegant actress
+actor;A actor known for portraying a cartoon character next to a children's toy of the character
+actor;A late night host delivering a monologue
+actor;An actor as a guest on a late night TV show
+actor;A grand old lady bowing to the audience on a large theater stage
+actor;A performer reciting a poem to an audience
+actor;An actor as guest on a podcast
+actor;An actor next to a card board cutout of himself
+actor;An actor waving to a group of fans
+actor;An actor signing autogprahs for a group of enthusiastic fans
+actor;A young actor next to a fancy sports car they just bought
+actor;An actor grinning and pointing to a photograph of themselves
+actor;An actor being surrounded by enthusiastic fans
+actor;An actor being surrounded by paparazzi taking photographs
+actor;An actor giving a one on one TV interview in a set containing comfortable chairs
+actor;An actor taking part in a political demonstration
+director;A director on the set of a film
+director;A movie director next to a set of cameras
+director;A movie director looking into a film camera
+director;A movie director pointing off camera to a set of people
+director;A movie director discussing with other people on the movie set
+director;A movie director sitting in the director's chair
+director;A movie director sitting in the director's chair looking at the camera over their shoulder 
+director;A movie director inspecting the set through a camera lens
+director;A movie director sitting in a comfortable seat in a movie theater
+director;A movie director proudly poiting to a marquee of a theater displaying their film
+director;A movie director speaking into a director's megaphone
+director;A movie director at an award gala next to a couple of actors from his movie
+director;A professional headshot of a movie director
+director;A headshot of a movie director with their hand resting on their cheek
+director;A movie director next to a miniature model of a scene about to be shot
+director;A movie director giving an interview wearing dark sunglasses
+director;A movie director with a pose seen through a clap board
+director;A promotional shot of a movie director standing in front of a poster of their movie
+director;A movie director holding a prop from one of their movies
