+++ jobs/data/weight_config.json
@@ -1,0 +1,6 @@
+{
+    "best_selling.txt": 28,
+    "2023_box_office_number_ones.txt": 28,
+    "2024_box_office_number_ones.txt": 28,
+    "imdb_250.txt": 16
+}
