+++ tests/test_app.py
@@ -27,16 +27,26 @@ def test_index_request(client):
         assert response.status_code == 200
         assert b"patched-render" in response.data
 
-def test_person_generation_request(client):
-    """Test content generation request for people."""
-    with patch("webserver.get_person_info.batch_translate_and_upload") as mock_batch_translate_and_upload:
-        response = client.get("/_generate", query_string={"type": "PERSON"}, headers={"X-Appengine-Cron": "1"})
-        mock_batch_translate_and_upload.assert_called()
+def test_fetch_movie_description_with_path(client):
+    """Should call download_description when path is provided."""
+    with patch("webserver.views.gcs_utils.download_description") as mock_download, \
+         patch("webserver.views.utils.format_as_html") as mock_format:
+        mock_download.return_value = "desc-data"
+        mock_format.return_value = "formatted-data"
+        response = client.get("/_get?path=some/path.txt")
+        mock_download.assert_called_once_with("some/path.txt")
+        mock_format.assert_called_once_with("desc-data")
         assert response.status_code == 200
+        assert b"formatted-data" in response.data
 
-def test_movie_generation_request(client):
-    """Test content generation request for movies."""
-    with patch("webserver.translate.batch_translate_and_upload") as mock_batch_translate_and_upload:
-        response = client.get("/_generate", query_string={"type": "MOVIE"}, headers={"X-Appengine-Cron": "1"})
-        mock_batch_translate_and_upload.assert_called()
+def test_fetch_movie_description_random(client):
+    """Should call download_random_content when no path is provided."""
+    with patch("webserver.views.gcs_utils.download_random_content") as mock_random, \
+         patch("webserver.views.utils.format_as_html") as mock_format:
+        mock_random.return_value = "random-data"
+        mock_format.return_value = "formatted-random"
+        response = client.get("/_get")
+        mock_random.assert_called_once()
+        mock_format.assert_called_once_with("random-data")
         assert response.status_code == 200
+        assert b"formatted-random" in response.data
